Fix content.opf
I was unable to open the minimal epub file in Apple Books (2.4) until I made this fix.

```
$ xmllint OEBPS/content.opf                                                                                            
OEBPS/content.opf:6: namespace error : Namespace prefix opf for role on creator is not defined                         
            <dc:creator opf:role="aut">
```

Namespace prefix `opf` is not defined. The namespace, `http://www.idpf.org/2007/opf`, is already set as the default namespace, so the prefix can simply be removed.